Fix tracking config for Blender 4.4 API changes
